Bugfixes, SQLite support, better terminal output, split script in 2. See README for details.
