+++ config.ini
@@ -1,0 +1,24 @@
+[database]
+engine = mysql
+host = localhost
+database = retrosheet
+schema = retrosheet
+user = user
+password = password
+
+[download]
+directory = files
+num_threads = 20
+
+[processing]
+num_threads = 20
+
+[chadwick]
+directory = /usr/local/bin/
+
+# Don't change this unless you know what you're doing
+[retrosheet]
+url = http://www.retrosheet.org/game.htm
+
+[debug]
+verbose = True
