--- db.ini
@@ -1,7 +1,0 @@
-[database]
-engine = postgresql
-host = localhost
-database = retrosheet
-schema = retrosheet
-;user = none
-;password = none
